requirements.txt
fastai2==0.0.16
fastcore==0.1.16
fastprogress==0.2.3
joblib==0.14.1
numpy==1.18.2
packaging==20.3
pandas==1.0.3
Pillow==7.1.1
streamlit==0.57.3
torch==1.4.0
torchvision==0.5.0

